refactor(datepicker): set up date range picker classes (#18292)
Moves some things around to set up the date range picker. Includes:
* Moves all of the datepicker logic into a base class so that it can be reused between the single and range selection.
* Makes the input being passed into a datepicker generic so that we can strongly type `MatDatepicker` to `MatDatepickerInput` and `MatDateRangePicker` to `MatDateRangeInput`.
* Sets up min, max and date filter validation for the date range picker.
* Hooks up the date range picker to the date range input.
* Makes it so that a generic datepicker can be passed to a datepicker toggle.
* Adds a `DateFilterFn` type so we don't have to repeat the signature for the filter function everywhere.

Note that while picking a date range somewhat works now, it's still a bit broken. The primary goal of these changes was to have all of components in place and wired up so that we can start working on the UI in follow-up PRs.
