+++ src/dev-app/datepicker/datepicker-demo.html
@@ -175,11 +175,12 @@
 <div class="demo-range-group">
   <mat-form-field>
     <mat-label>Enter a date range</mat-label>
-    <mat-date-range-input [formGroup]="range1">
+    <mat-date-range-input [formGroup]="range1" [rangePicker]="range1Picker">
       <input matStartDate formControlName="start" placeholder="Start date"/>
       <input matEndDate formControlName="end" placeholder="End date"/>
     </mat-date-range-input>
-    <mat-datepicker-toggle matSuffix></mat-datepicker-toggle>
+    <mat-datepicker-toggle [for]="range1Picker" matSuffix></mat-datepicker-toggle>
+    <mat-date-range-picker #range1Picker></mat-date-range-picker>
   </mat-form-field>
   <div>{{range1.value | json}}</div>
 </div>
@@ -187,11 +188,12 @@
 <div class="demo-range-group">
   <mat-form-field appearance="fill">
     <mat-label>Enter a date range</mat-label>
-    <mat-date-range-input [formGroup]="range2">
+    <mat-date-range-input [formGroup]="range2" [rangePicker]="range2Picker">
       <input matStartDate formControlName="start" placeholder="Start date"/>
       <input matEndDate formControlName="end" placeholder="End date"/>
     </mat-date-range-input>
-    <mat-datepicker-toggle matSuffix></mat-datepicker-toggle>
+    <mat-datepicker-toggle [for]="range2Picker" matSuffix></mat-datepicker-toggle>
+    <mat-date-range-picker #range2Picker></mat-date-range-picker>
   </mat-form-field>
   <div>{{range2.value | json}}</div>
 </div>
@@ -199,11 +201,12 @@
 <div class="demo-range-group">
   <mat-form-field appearance="outline">
     <mat-label>Enter a date range</mat-label>
-    <mat-date-range-input [formGroup]="range3">
+    <mat-date-range-input [formGroup]="range3" [rangePicker]="range3Picker">
       <input matStartDate formControlName="start" placeholder="Start date"/>
       <input matEndDate formControlName="end" placeholder="End date"/>
     </mat-date-range-input>
-    <mat-datepicker-toggle matSuffix></mat-datepicker-toggle>
+    <mat-datepicker-toggle [for]="range3Picker" matSuffix></mat-datepicker-toggle>
+    <mat-date-range-picker #range3Picker></mat-date-range-picker>
   </mat-form-field>
   <div>{{range3.value | json}}</div>
 </div>
