+++ src/material/datepicker/datepicker-base.ts
@@ -1,0 +1,541 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
+import {Directionality} from '@angular/cdk/bidi';
+import {BooleanInput, coerceBooleanProperty} from '@angular/cdk/coercion';
+import {ESCAPE, UP_ARROW} from '@angular/cdk/keycodes';
+import {
+  Overlay,
+  OverlayConfig,
+  OverlayRef,
+  PositionStrategy,
+  ScrollStrategy,
+} from '@angular/cdk/overlay';
+import {ComponentPortal, ComponentType} from '@angular/cdk/portal';
+import {DOCUMENT} from '@angular/common';
+import {
+  AfterViewInit,
+  ChangeDetectionStrategy,
+  Component,
+  ComponentRef,
+  ElementRef,
+  EventEmitter,
+  Inject,
+  InjectionToken,
+  Input,
+  NgZone,
+  OnDestroy,
+  Optional,
+  Output,
+  ViewChild,
+  ViewContainerRef,
+  ViewEncapsulation,
+  ChangeDetectorRef,
+  Directive,
+} from '@angular/core';
+import {
+  CanColor,
+  CanColorCtor,
+  DateAdapter,
+  mixinColor,
+  ThemePalette,
+  MatDateSelectionModel,
+  ExtractDateTypeFromSelection,
+} from '@angular/material/core';
+import {MatDialog, MatDialogRef} from '@angular/material/dialog';
+import {merge, Subject, Observable} from 'rxjs';
+import {filter, take} from 'rxjs/operators';
+import {MatCalendar} from './calendar';
+import {matDatepickerAnimations} from './datepicker-animations';
+import {createMissingDateImplError} from './datepicker-errors';
+import {MatCalendarCellCssClasses} from './calendar-body';
+import {DateFilterFn} from './datepicker-input-base';
+
+/** Used to generate a unique ID for each datepicker instance. */
+let datepickerUid = 0;
+
+/** Injection token that determines the scroll handling while the calendar is open. */
+export const MAT_DATEPICKER_SCROLL_STRATEGY =
+    new InjectionToken<() => ScrollStrategy>('mat-datepicker-scroll-strategy');
+
+/** @docs-private */
+export function MAT_DATEPICKER_SCROLL_STRATEGY_FACTORY(overlay: Overlay): () => ScrollStrategy {
+  return () => overlay.scrollStrategies.reposition();
+}
+
+/** @docs-private */
+export const MAT_DATEPICKER_SCROLL_STRATEGY_FACTORY_PROVIDER = {
+  provide: MAT_DATEPICKER_SCROLL_STRATEGY,
+  deps: [Overlay],
+  useFactory: MAT_DATEPICKER_SCROLL_STRATEGY_FACTORY,
+};
+
+// Boilerplate for applying mixins to MatDatepickerContent.
+/** @docs-private */
+class MatDatepickerContentBase {
+  constructor(public _elementRef: ElementRef) { }
+}
+const _MatDatepickerContentMixinBase: CanColorCtor & typeof MatDatepickerContentBase =
+    mixinColor(MatDatepickerContentBase);
+
+/**
+ * Component used as the content for the datepicker dialog and popup. We use this instead of using
+ * MatCalendar directly as the content so we can control the initial focus. This also gives us a
+ * place to put additional features of the popup that are not part of the calendar itself in the
+ * future. (e.g. confirmation buttons).
+ * @docs-private
+ */
+@Component({
+  selector: 'mat-datepicker-content',
+  templateUrl: 'datepicker-content.html',
+  styleUrls: ['datepicker-content.css'],
+  host: {
+    'class': 'mat-datepicker-content',
+    '[@transformPanel]': '_animationState',
+    '(@transformPanel.done)': '_animationDone.next()',
+    '[class.mat-datepicker-content-touch]': 'datepicker.touchUi',
+  },
+  animations: [
+    matDatepickerAnimations.transformPanel,
+    matDatepickerAnimations.fadeInCalendar,
+  ],
+  exportAs: 'matDatepickerContent',
+  encapsulation: ViewEncapsulation.None,
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  inputs: ['color'],
+})
+export class MatDatepickerContent<S, D = ExtractDateTypeFromSelection<S>>
+  extends _MatDatepickerContentMixinBase implements AfterViewInit, OnDestroy, CanColor {
+
+  /** Reference to the internal calendar component. */
+  @ViewChild(MatCalendar) _calendar: MatCalendar<D>;
+
+  /** Reference to the datepicker that created the overlay. */
+  datepicker: MatDatepickerBase<any, S, D>;
+
+  /** Whether the datepicker is above or below the input. */
+  _isAbove: boolean;
+
+  /** Current state of the animation. */
+  _animationState: 'enter' | 'void' = 'enter';
+
+  /** Emits when an animation has finished. */
+  _animationDone = new Subject<void>();
+
+  constructor(
+    elementRef: ElementRef,
+    /**
+     * @deprecated `_changeDetectorRef` parameter to become required.
+     * @breaking-change 11.0.0
+     */
+    private _changeDetectorRef?: ChangeDetectorRef) {
+    super(elementRef);
+  }
+
+  ngAfterViewInit() {
+    this._calendar.focusActiveCell();
+  }
+
+  ngOnDestroy() {
+    this._animationDone.complete();
+  }
+
+  _startExitAnimation() {
+    this._animationState = 'void';
+
+    // @breaking-change 11.0.0 Remove null check for `_changeDetectorRef`.
+    if (this._changeDetectorRef) {
+      this._changeDetectorRef.markForCheck();
+    }
+  }
+}
+
+/** Form control that can be associated with a datepicker. */
+export interface MatDatepickerControl<D> {
+  getStartValue(): D | null;
+  getThemePalette(): ThemePalette;
+  min: D | null;
+  max: D | null;
+  disabled: boolean;
+  dateFilter: DateFilterFn<D>;
+  getConnectedOverlayOrigin(): ElementRef;
+  _disabledChange: Observable<boolean>;
+}
+
+/** Base class for a datepicker. */
+@Directive()
+export abstract class MatDatepickerBase<C extends MatDatepickerControl<D>, S,
+  D = ExtractDateTypeFromSelection<S>> implements OnDestroy, CanColor {
+  private _scrollStrategy: () => ScrollStrategy;
+
+  /** An input indicating the type of the custom header component for the calendar, if set. */
+  @Input() calendarHeaderComponent: ComponentType<any>;
+
+  /** The date to open the calendar to initially. */
+  @Input()
+  get startAt(): D | null {
+    // If an explicit startAt is set we start there, otherwise we start at whatever the currently
+    // selected value is.
+    return this._startAt || (this._datepickerInput ? this._datepickerInput.getStartValue() : null);
+  }
+  set startAt(value: D | null) {
+    this._startAt = this._getValidDateOrNull(this._dateAdapter.deserialize(value));
+  }
+  private _startAt: D | null;
+
+  /** The view that the calendar should start in. */
+  @Input() startView: 'month' | 'year' | 'multi-year' = 'month';
+
+  /** Color palette to use on the datepicker's calendar. */
+  @Input()
+  get color(): ThemePalette {
+    return this._color ||
+        (this._datepickerInput ? this._datepickerInput.getThemePalette() : undefined);
+  }
+  set color(value: ThemePalette) {
+    this._color = value;
+  }
+  _color: ThemePalette;
+
+  /**
+   * Whether the calendar UI is in touch mode. In touch mode the calendar opens in a dialog rather
+   * than a popup and elements have more padding to allow for bigger touch targets.
+   */
+  @Input()
+  get touchUi(): boolean { return this._touchUi; }
+  set touchUi(value: boolean) {
+    this._touchUi = coerceBooleanProperty(value);
+  }
+  private _touchUi = false;
+
+  /** Whether the datepicker pop-up should be disabled. */
+  @Input()
+  get disabled(): boolean {
+    return this._disabled === undefined && this._datepickerInput ?
+        this._datepickerInput.disabled : !!this._disabled;
+  }
+  set disabled(value: boolean) {
+    const newValue = coerceBooleanProperty(value);
+
+    if (newValue !== this._disabled) {
+      this._disabled = newValue;
+      this._disabledChange.next(newValue);
+    }
+  }
+  private _disabled: boolean;
+
+  /**
+   * Emits selected year in multiyear view.
+   * This doesn't imply a change on the selected date.
+   */
+  @Output() readonly yearSelected: EventEmitter<D> = new EventEmitter<D>();
+
+  /**
+   * Emits selected month in year view.
+   * This doesn't imply a change on the selected date.
+   */
+  @Output() readonly monthSelected: EventEmitter<D> = new EventEmitter<D>();
+
+  /** Classes to be passed to the date picker panel. Supports the same syntax as `ngClass`. */
+  @Input() panelClass: string | string[];
+
+  /** Function that can be used to add custom CSS classes to dates. */
+  @Input() dateClass: (date: D) => MatCalendarCellCssClasses;
+
+  /** Emits when the datepicker has been opened. */
+  @Output('opened') openedStream: EventEmitter<void> = new EventEmitter<void>();
+
+  /** Emits when the datepicker has been closed. */
+  @Output('closed') closedStream: EventEmitter<void> = new EventEmitter<void>();
+
+
+  /** Whether the calendar is open. */
+  @Input()
+  get opened(): boolean { return this._opened; }
+  set opened(value: boolean) { value ? this.open() : this.close(); }
+  private _opened = false;
+
+  /** The id for the datepicker calendar. */
+  id: string = `mat-datepicker-${datepickerUid++}`;
+
+  /** The minimum selectable date. */
+  get _minDate(): D | null {
+    return this._datepickerInput && this._datepickerInput.min;
+  }
+
+  /** The maximum selectable date. */
+  get _maxDate(): D | null {
+    return this._datepickerInput && this._datepickerInput.max;
+  }
+
+  get _dateFilter(): DateFilterFn<D> {
+    return this._datepickerInput && this._datepickerInput.dateFilter;
+  }
+
+  /** A reference to the overlay when the calendar is opened as a popup. */
+  private _popupRef: OverlayRef | null;
+
+  /** A reference to the dialog when the calendar is opened as a dialog. */
+  private _dialogRef: MatDialogRef<MatDatepickerContent<S, D>> | null;
+
+  /** Reference to the component instantiated in popup mode. */
+  private _popupComponentRef: ComponentRef<MatDatepickerContent<S, D>> | null;
+
+  /** The element that was focused before the datepicker was opened. */
+  private _focusedElementBeforeOpen: HTMLElement | null = null;
+
+  /** The input element this datepicker is associated with. */
+  _datepickerInput: C;
+
+  /** Emits when the datepicker is disabled. */
+  readonly _disabledChange = new Subject<boolean>();
+
+  constructor(private _dialog: MatDialog,
+              private _overlay: Overlay,
+              private _ngZone: NgZone,
+              private _viewContainerRef: ViewContainerRef,
+              @Inject(MAT_DATEPICKER_SCROLL_STRATEGY) scrollStrategy: any,
+              @Optional() private _dateAdapter: DateAdapter<D>,
+              @Optional() private _dir: Directionality,
+              @Optional() @Inject(DOCUMENT) private _document: any,
+              private _model: MatDateSelectionModel<S, D>) {
+    if (!this._dateAdapter) {
+      throw createMissingDateImplError('DateAdapter');
+    }
+
+    this._scrollStrategy = scrollStrategy;
+  }
+
+  ngOnDestroy() {
+    this._destroyPopup();
+    this.close();
+    this._disabledChange.complete();
+  }
+
+  /** Selects the given date */
+  select(date: D): void {
+    this._model.add(date);
+  }
+
+  /** Emits the selected year in multiyear view */
+  _selectYear(normalizedYear: D): void {
+    this.yearSelected.emit(normalizedYear);
+  }
+
+  /** Emits selected month in year view */
+  _selectMonth(normalizedMonth: D): void {
+    this.monthSelected.emit(normalizedMonth);
+  }
+
+  /**
+   * Register an input with this datepicker.
+   * @param input The datepicker input to register with this datepicker.
+   * @returns Selection model that the input should hook itself up to.
+   */
+  _registerInput(input: C): MatDateSelectionModel<S, D> {
+    if (this._datepickerInput) {
+      throw Error('A MatDatepicker can only be associated with a single input.');
+    }
+    this._datepickerInput = input;
+    return this._model;
+  }
+
+  /** Open the calendar. */
+  open(): void {
+    if (this._opened || this.disabled) {
+      return;
+    }
+    if (!this._datepickerInput) {
+      throw Error('Attempted to open an MatDatepicker with no associated input.');
+    }
+    if (this._document) {
+      this._focusedElementBeforeOpen = this._document.activeElement;
+    }
+
+    this.touchUi ? this._openAsDialog() : this._openAsPopup();
+    this._opened = true;
+    this.openedStream.emit();
+  }
+
+  /** Close the calendar. */
+  close(): void {
+    if (!this._opened) {
+      return;
+    }
+    if (this._popupComponentRef && this._popupRef) {
+      const instance = this._popupComponentRef.instance;
+      instance._startExitAnimation();
+      instance._animationDone.pipe(take(1)).subscribe(() => this._destroyPopup());
+    }
+    if (this._dialogRef) {
+      this._dialogRef.close();
+      this._dialogRef = null;
+    }
+
+    const completeClose = () => {
+      // The `_opened` could've been reset already if
+      // we got two events in quick succession.
+      if (this._opened) {
+        this._opened = false;
+        this.closedStream.emit();
+        this._focusedElementBeforeOpen = null;
+      }
+    };
+
+    if (this._focusedElementBeforeOpen &&
+      typeof this._focusedElementBeforeOpen.focus === 'function') {
+      // Because IE moves focus asynchronously, we can't count on it being restored before we've
+      // marked the datepicker as closed. If the event fires out of sequence and the element that
+      // we're refocusing opens the datepicker on focus, the user could be stuck with not being
+      // able to close the calendar at all. We work around it by making the logic, that marks
+      // the datepicker as closed, async as well.
+      this._focusedElementBeforeOpen.focus();
+      setTimeout(completeClose);
+    } else {
+      completeClose();
+    }
+  }
+
+  /** Open the calendar as a dialog. */
+  private _openAsDialog(): void {
+    // Usually this would be handled by `open` which ensures that we can only have one overlay
+    // open at a time, however since we reset the variables in async handlers some overlays
+    // may slip through if the user opens and closes multiple times in quick succession (e.g.
+    // by holding down the enter key).
+    if (this._dialogRef) {
+      this._dialogRef.close();
+    }
+
+    this._dialogRef = this._dialog.open<MatDatepickerContent<S, D>>(MatDatepickerContent, {
+      direction: this._dir ? this._dir.value : 'ltr',
+      viewContainerRef: this._viewContainerRef,
+      panelClass: 'mat-datepicker-dialog',
+
+      // These values are all the same as the defaults, but we set them explicitly so that the
+      // datepicker dialog behaves consistently even if the user changed the defaults.
+      hasBackdrop: true,
+      disableClose: false,
+      width: '',
+      height: '',
+      minWidth: '',
+      minHeight: '',
+      maxWidth: '80vw',
+      maxHeight: '',
+      position: {},
+      autoFocus: true,
+      restoreFocus: true
+    });
+
+    this._dialogRef.afterClosed().subscribe(() => this.close());
+    this._dialogRef.componentInstance.datepicker = this;
+    this._dialogRef.componentInstance.color = this.color;
+  }
+
+  /** Open the calendar as a popup. */
+  private _openAsPopup(): void {
+    const portal = new ComponentPortal<MatDatepickerContent<S, D>>(MatDatepickerContent,
+                                                                   this._viewContainerRef);
+
+    this._destroyPopup();
+    this._createPopup();
+    const ref = this._popupComponentRef = this._popupRef!.attach(portal);
+    ref.instance.datepicker = this;
+    ref.instance.color = this.color;
+
+    // Update the position once the calendar has rendered.
+    this._ngZone.onStable.asObservable().pipe(take(1)).subscribe(() => {
+      this._popupRef!.updatePosition();
+    });
+  }
+
+  /** Create the popup. */
+  private _createPopup(): void {
+    const overlayConfig = new OverlayConfig({
+      positionStrategy: this._createPopupPositionStrategy(),
+      hasBackdrop: true,
+      backdropClass: 'mat-overlay-transparent-backdrop',
+      direction: this._dir,
+      scrollStrategy: this._scrollStrategy(),
+      panelClass: 'mat-datepicker-popup',
+    });
+
+    this._popupRef = this._overlay.create(overlayConfig);
+    this._popupRef.overlayElement.setAttribute('role', 'dialog');
+
+    merge(
+      this._popupRef.backdropClick(),
+      this._popupRef.detachments(),
+      this._popupRef.keydownEvents().pipe(filter(event => {
+        // Closing on alt + up is only valid when there's an input associated with the datepicker.
+        return event.keyCode === ESCAPE ||
+               (this._datepickerInput && event.altKey && event.keyCode === UP_ARROW);
+      }))
+    ).subscribe(event => {
+      if (event) {
+        event.preventDefault();
+      }
+
+      this.close();
+    });
+  }
+
+  /** Destroys the current popup overlay. */
+  private _destroyPopup() {
+    if (this._popupRef) {
+      this._popupRef.dispose();
+      this._popupRef = this._popupComponentRef = null;
+    }
+  }
+
+  /** Create the popup PositionStrategy. */
+  private _createPopupPositionStrategy(): PositionStrategy {
+    return this._overlay.position()
+      .flexibleConnectedTo(this._datepickerInput.getConnectedOverlayOrigin())
+      .withTransformOriginOn('.mat-datepicker-content')
+      .withFlexibleDimensions(false)
+      .withViewportMargin(8)
+      .withLockedPosition()
+      .withPositions([
+        {
+          originX: 'start',
+          originY: 'bottom',
+          overlayX: 'start',
+          overlayY: 'top'
+        },
+        {
+          originX: 'start',
+          originY: 'top',
+          overlayX: 'start',
+          overlayY: 'bottom'
+        },
+        {
+          originX: 'end',
+          originY: 'bottom',
+          overlayX: 'end',
+          overlayY: 'top'
+        },
+        {
+          originX: 'end',
+          originY: 'top',
+          overlayX: 'end',
+          overlayY: 'bottom'
+        }
+      ]);
+  }
+
+  /**
+   * @param obj The object to check.
+   * @returns The given object if it is both a date instance and valid, otherwise null.
+   */
+  private _getValidDateOrNull(obj: any): D | null {
+    return (this._dateAdapter.isDateInstance(obj) && this._dateAdapter.isValid(obj)) ? obj : null;
+  }
+
+  static ngAcceptInputType_disabled: BooleanInput;
+  static ngAcceptInputType_touchUi: BooleanInput;
+}
